move ErgoTool into appkit.examples package
