+++ scala-examples/src/main/java/org/ergoplatform/appkit/examples/ErgoTool.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.example
+package org.ergoplatform.appkit.examples
 
 import java.io.PrintStream
 import java.util.Arrays
