+++ scala-examples/src/main/java/org/ergoplatform/appkit/examples/RunMockedScala.scala
@@ -1,9 +1,9 @@
-package org.ergoplatform.example
+package org.ergoplatform.appkit.examples
 
-import org.ergoplatform.example.util.FileMockedErgoClient
+import org.ergoplatform.appkit.examples.util.FileMockedErgoClient
 
 object RunMockedScala extends App {
-  import org.ergoplatform.example.MockData._
+  import org.ergoplatform.appkit.examples.MockData._
   val res = new FileMockedErgoClient(infoFile, lastHeadersFile, boxFile).execute { ctx =>
     val r = new ExampleScenarios(ctx)
     val res = r.aggregateUtxoBoxes("abc", 10, "83b94f2df7e97586a9fe8fe43fa84d252aa74ecee5fe0871f85a45663927cd9a")
